Migrate Salesloft MCP server to static metadata pattern
- Move server from mcp_internal_actions/servers to api/actions/servers
- Create metadata.ts with SALESLOFT_TOOLS_METADATA using createToolsRecord
- Create types.ts with Salesloft API types
- Create client.ts with API request helpers
- Create helpers.ts with token retrieval and action formatting
- Create tools/index.ts with typed handlers using buildTools
- Update constants.ts to use SALESLOFT_SERVER metadata import
- Update servers/index.ts to import from new location
- Delete old salesloft server files
